fix(location-geoclue2): use gdouble return type
The function g_variant_get_double should have a gdouble return type.
glib likes to make up types to replace perfectly good C types.

+++ src/location-geoclue2.c
@@ -36,8 +36,8 @@ typedef struct {
 	int pipe_fd_write;
 	int available;
 	int error;
-	float latitude;
-	float longitude;
+	gdouble latitude;
+	gdouble longitude;
 } location_geoclue2_state_t;
